Chatbot - Copy (2)/requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-huggingface
clean-text
chromadb
pypdf
python-dotenv
pandas
sentence-transformers
langchain-text-splitters
